requirements.txt
pyserial
flask
jsonify
requests
